requirements.txt
aiohttp==3.8.4
aioimaplib @ git+https://github.com/bamthomas/aioimaplib@25887f7a0f03797f02c9336c168fb11a503650d0
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
discord==2.2.3
discord.py==2.2.3
frozenlist==1.3.3
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
httpx-oauth==0.11.2
idna==3.4
IMAPClient==2.3.1
imgkit==1.2.3
multidict==6.0.4
nest-asyncio==1.5.6
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
urllib3==1.26.16
yarl==1.9.2
